Improve html export
