+++ package.json
@@ -42,6 +42,7 @@
     "webpack-dev-server": "2.1.0-beta.12"
   },
   "dependencies": {
+    "common-tags": "^1.4.0",
     "draft-convert": "^1.3.1",
     "draft-js": "^0.9.1",
     "draft-js-export-html": "^0.5.2",
