+++ src/plugins/video/VideoBlock.js
@@ -21,44 +21,11 @@ export default class extends Component {
     this.props.container.updateData({caption: event.target.value})
   }
 
-  getVideoIdYoutube(str) {
-    // link : https://youtube.com/watch?v=HBHJ0XGZfLs
-    // share : https://youtu.be/HBHJ0XGZfLs
-    // embed : https://youtube.com/embed/HBHJ0XGZfLs
-    var re = /\/\/(?:www\.)?youtu(?:\.be|be\.com)\/(?:watch\?v=|embed\/)?([a-z0-9_\-]+)/i
-    var matches = re.exec(str)
-    return matches && matches[1]
-  }
-
-  getVideoIdVimeo(str) {
-    // embed & link: https://vimeo.com/713300
-    var re = /\/\/(?:www\.)?vimeo.com\/([0-9a-z\-_]+)/i
-    var matches = re.exec(str)
-    return matches && matches[1]
-  }
-
-  getVideoUrl(src) {
-    /* youtube */
-    let id = this.getVideoIdYoutube(src)
-    if (id !== null) {
-      return `https://youtube.com/embed/${id}`
-    }
-
-    /* vimeo */
-    id = this.getVideoIdVimeo(src)
-    if (id !== null) {
-      return `https://player.vimeo.com/video/${id}`
-    }
-  }
-
   render () {
-    let videoSrc = this.getVideoUrl(this.props.data.src)
-    if(videoSrc === undefined) { return null }
-
     return (
       <Block {...this.props} actions={this.actions}>
         <VideoBlockWrapper className='ld-video-block-wrapper'>
-          <VideoBlock src={videoSrc}
+          <VideoBlock src={this.props.data.src}
                   className='ld-video-block'
                   frameBorder='0'
                   allowFullScreen />
